不要なもの削除と細かな修正

+++ GameTemplate/Game/Skill/SkillHeal.cpp
@@ -109,7 +109,7 @@ int SkillHeal::WideHeal()
 	int res = 0;
 	for (int i = 0; i < list.size(); i++) {
 		//全体にバフをかける。
-		res += list[i]->Healing(result);
+		res += list[i]->Healing(static_cast<int>(result));
 	}
 	res /= static_cast<int>(list.size());		//上昇値の平均をとる。
